fixed db connection and security bugs

+++ WasteLess/src/main/java/com/assignment1/wasteless/Data/Repository/GroceryListItemRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface GroceryListItemRepository extends JpaRepository<GroceryListItem, Long> {
+public interface GroceryListItemRepository extends JpaRepository<GroceryListItem, Integer> {
 }
